Mapper fra vedleggRequest til SafRequest for at navn på variablene stemmer overens med graphql-spørring. Det er noe avvik i begrep-bruk mellom frontend ef-sak og saf api
